+++ app/src/main/res/layout/fragment_github_search.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TopActivity">
+    tools:context=".presentation.TopActivity">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/searchBar"
