+++ app/src/main/res/navigation/nav_graph.xml
@@ -18,7 +18,7 @@
 
     <fragment
         android:id="@+id/gitHubDiscriptionFragment"
-        android:name="jp.co.yumemi.android.code_check.GitHubDiscriptionFragment"
+        android:name="jp.co.yumemi.android.code_check.presentation.GitHubDiscriptionFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_github_discription">
         <argument
